backend/pyproject.toml
[tool.poetry]
name = "barilan-backend"
version = "0.1.0"
description = ""
authors = ["talabElad <talab@elad.co.il>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.35.76"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
